Clean up Moonshine STT runtime structure
