+++ include/engine/models/moonshine_stt/session.h
@@ -1,16 +1,15 @@
 #pragma once
 
 #include "engine/framework/runtime/session_base.h"
-#include "engine/framework/modules/activation_modules.h"
-#include "engine/models/moonshine_stt/assets.h"
-#include "engine/models/moonshine_stt/weights.h"
+#include "engine/models/moonshine_stt/runtime.h"
 
-#include <cstddef>
 #include <memory>
 #include <string>
-#include <unordered_map>
 
 namespace engine::models::moonshine_stt {
+
+struct MoonshineAssets;
+struct MoonshineWeights;
 
 class MoonshineSTTSession final
     : public runtime::RuntimeSessionBase
@@ -37,23 +36,13 @@ public:
     runtime::TaskResult finish_stream() override;
 
 private:
-    runtime::TaskResult transcribe(
-        const runtime::AudioBuffer & audio,
-        const std::unordered_map<std::string, std::string> & options);
-
     runtime::TaskSpec task_;
     std::shared_ptr<const MoonshineAssets> assets_;
+    MoonshineRuntimeConfig runtime_config_;
     std::shared_ptr<const MoonshineWeights> weights_;
     runtime::StreamEventCallback stream_event_sink_;
     runtime::AudioBuffer streaming_audio_;
     runtime::TaskRequest streaming_request_;
-    size_t weight_context_bytes_ = 256ull * 1024ull * 1024ull;
-    size_t graph_arena_bytes_ = 512ull * 1024ull * 1024ull;
-    engine::modules::GeluApproximation encoder_gelu_ = engine::modules::GeluApproximation::Quick;
-    engine::assets::TensorStorageType encoder_weight_storage_type_ = engine::assets::TensorStorageType::Native;
-    engine::assets::TensorStorageType decoder_weight_storage_type_ = engine::assets::TensorStorageType::Native;
-    engine::assets::TensorStorageType conv_weight_storage_type_ = engine::assets::TensorStorageType::Native;
-    bool cpu_blas_scheduler_ = true;
     bool stream_started_ = false;
 };
